delete ticket

+++ components/shared/DeleteOrder.tsx
@@ -1,26 +1,54 @@
-"use client"
+'use client'
 
 import { useTransition } from 'react'
 import { usePathname } from 'next/navigation'
-import { deleteOrder } from '@/lib/actions/order.actions'
-import { Button } from '../ui/button'
+import Image from 'next/image'
 
-export function DeleteOrder({ orderId }: { orderId: string }) {
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/components/ui/alert-dialog'
+
+import { deleteOrder } from '@/lib/actions/order.actions'
+
+export const DeleteOrder = ({ orderId }: { orderId: string }) => {
   const pathname = usePathname()
-  const [isPending, startTransition] = useTransition()
+  let [isPending, startTransition] = useTransition()
 
   return (
-    <Button
-      variant="destructive"
-      size="sm"
-      disabled={isPending}
-      onClick={() => {
-        startTransition(async () => {
-          await deleteOrder(orderId)
-        })
-      }}
-    >
-      {isPending ? 'Deleting...' : 'Delete'}
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger>
+        <Image src="/assets/icons/delete.svg" alt="delete" width={20} height={20} />
+      </AlertDialogTrigger>
+
+      <AlertDialogContent className="bg-white">
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you sure you want to delete this ticket?</AlertDialogTitle>
+          <AlertDialogDescription className="p-regular-16 text-grey-600">
+            This will permanently delete this ticket
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+
+          <AlertDialogAction
+            onClick={() =>
+              startTransition(async () => {
+                await deleteOrder(orderId)
+              })
+            }>
+            {isPending ? 'Deleting...' : 'Delete'}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   )
 }
